Use flexbox to style webpage layout
Use flexbox in body to separate header, content, and footer using
space-between. Use left-links class to have links be next to each other
and gap to add space between. Similar with right-links but adding
margin-left auto to move to the right. Use flex for
footer and header to flex left-links and right-links for margin-left
auto to work. Add background-color to footer and removed
default text styles and decoration for lists and links.